Move Micro loader classes to loader module
