+++ arquillian/server/pom.xml
@@ -49,7 +49,7 @@
 
         <dependency>
             <groupId>cloud.piranha</groupId>
-            <artifactId>piranha-micro</artifactId>
+            <artifactId>piranha-micro-loader</artifactId>
             <version>${project.version}</version>
         </dependency>
