+++ arquillian/server/src/main/java/cloud/piranha/arquillian/server/PiranhaServerLoadableExtension.java
@@ -36,7 +36,7 @@ import org.jboss.arquillian.container.spi.client.protocol.ProtocolDescription;
 import org.jboss.arquillian.core.spi.LoadableExtension;
 import org.jboss.shrinkwrap.descriptor.api.Descriptor;
 
-import cloud.piranha.micro.MicroConfiguration;
+import cloud.piranha.micro.loader.MicroConfiguration;
 
 /**
  * The extension sets up the Arquillian Server Connector
