+++ micro/builder/pom.xml
@@ -13,7 +13,7 @@
     <dependencies>
         <dependency>
             <groupId>cloud.piranha</groupId>
-            <artifactId>piranha-micro</artifactId>
+            <artifactId>piranha-micro-loader</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
