+++ micro/builder/src/main/java/module-info.java
@@ -37,7 +37,7 @@ module cloud.piranha.micro.embedded {
     
     opens cloud.piranha.micro.embedded;
     
-    requires cloud.piranha.micro;
+    requires cloud.piranha.micro.loader;
     requires cloud.piranha.api;
     requires cloud.piranha.naming.api;
     requires cloud.piranha.naming.impl;
