+++ micro/distribution/src/main/javadoc/cloud.piranha.micro/cloud/piranha/micro/doc-files/request-response.png
