+++ micro/loader/pom.xml
@@ -7,18 +7,9 @@
         <artifactId>project</artifactId>
         <version>21.2.0-SNAPSHOT</version>
     </parent>
-    <!--
-
-      Note this is diverging from the established POM directory structure
-      because we needed a directory structure for the various Piranha Micro
-      Maven modules, but we already where using the micro directory for the
-      Piranha Micro distribution. This puts the distribution in the new micro
-      directory structure as child module, but retains its original Maven
-      coordinates.
-
-      -->
+   
     <groupId>cloud.piranha</groupId>
-    <artifactId>piranha-micro</artifactId>
+    <artifactId>piranha-micro-loader</artifactId>
     <packaging>jar</packaging>
     <name>Piranha - Micro - Runtime</name>
     <dependencies>
