+++ micro/loader/src/main/java/cloud/piranha/micro/loader/MicroConfiguration.java
@@ -25,7 +25,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package cloud.piranha.micro;
+package cloud.piranha.micro.loader;
 
 import static java.util.Arrays.stream;
 import static java.util.stream.Collectors.toList;
