+++ micro/loader/src/main/java/cloud/piranha/micro/loader/MicroDeployOutcome.java
@@ -25,7 +25,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package cloud.piranha.micro;
+package cloud.piranha.micro.loader;
 
 import java.util.Map;
 import java.util.Set;
