+++ micro/loader/src/main/java/cloud/piranha/micro/loader/StaticStreamHandler.java
@@ -1,0 +1,70 @@
+/*
+ * Copyright (c) 2002-2021 Manorrock.com. All Rights Reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without 
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *   1. Redistributions of source code must retain the above copyright notice, 
+ *      this list of conditions and the following disclaimer.
+ *   2. Redistributions in binary form must reproduce the above copyright
+ *      notice, this list of conditions and the following disclaimer in the
+ *      documentation and/or other materials provided with the distribution.
+ *   3. Neither the name of the copyright holder nor the names of its 
+ *      contributors may be used to endorse or promote products derived from
+ *      this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE 
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR 
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF 
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN 
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+package cloud.piranha.micro.loader;
+
+import java.io.IOException;
+import java.net.URL;
+import java.net.URLConnection;
+import java.net.URLStreamHandler;
+import java.util.Map;
+import java.util.function.Function;
+
+/**
+ * A static URL stream handler.
+ * 
+ * @author Manfred Riem (mriem@manorrock.com)
+ */
+public class StaticStreamHandler extends URLStreamHandler {
+    
+    /**
+     * Stores the protocol.
+     */
+    private final String protocol;
+    
+    /**
+     * Stores the handlers.
+     */
+    private final Map<String, Function<URL, URLConnection>> handlers;
+    
+    /**
+     * Constructor.
+     * 
+     * @param protocol the protocol.
+     * @param handlers the handlers.
+     */
+    public StaticStreamHandler(String protocol, Map<String, Function<URL, URLConnection>> handlers) {
+        this.protocol = protocol;
+        this.handlers = handlers;
+    }
+
+    @Override
+    protected URLConnection openConnection(URL u) throws IOException {
+        return handlers.get(protocol).apply(u);
+    }
+    
+}
